% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-s3.R
\name{aggregate.model_imp_tbl}
\alias{aggregate.model_imp_tbl}
\title{Aggregate model importance scores across tasks to compute overall importance
for each model}
\usage{
\method{aggregate}{model_imp_tbl}(
  x,
  by = "model_id",
  na_action = c("drop", "worst", "average"),
  fun = mean,
  ...
)
}
\arguments{
\item{x}{An object of class \code{model_imp_tbl}.}

\item{by}{A character vector with column names specifying the grouping
variable(s) for summarization. Default is \code{"model_id"}, which summarizes
importance scores for each model across all tasks.}

\item{na_action}{A character string specifying how to handle \code{NA} values
generated during importance score calculation for each task, occurring when a
model did not contribute to the ensemble prediction for a given task by
missing its forecast submission.
Three options are available: \code{c("drop", "worst", "average")}.
For each specific prediction task, each option works as follows:
\itemize{
\item \code{"drop"} removes \code{NA}s.
\item \code{"worst"} replaces \code{NA}s with the smallest value among importance metrics
of the other models.
\item \code{"average"} replaces \code{NA}s with the average value from the other
models' importance metrics.
}}

\item{fun}{A function used to summarize importance scores.
Default is \code{mean()}}

\item{...}{Additional arguments passed to the summary function \code{fun}.
(e.g., \verb{fun = quantile, probs = 0.25} for a quartile summary)}
}
\value{
A data frame with columns \code{model_id} and \verb{importance_score_<fun>},
where \verb{<fun>} is the name of the summary function
used (e.g., \code{importance_score_mean} when \code{fun = mean}).
The output is sorted in descending order of the summary importance scores.
}
\description{
Aggregate model importance scores across tasks to compute overall importance
for each model
}
\details{
This method extends \code{stats::aggregate} for objects of class \code{model_imp_tbl}.
}
