% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_metric_selection.R
\name{check_metric_selection}
\alias{check_metric_selection}
\title{Validate metric selection for output types}
\usage{
check_metric_selection(output_type, scoring_rule)
}
\arguments{
\item{output_type}{A character string specifying the type of prediction
output. Should be one of "mean", "median", "quantile", "pmf", or "sample".}

\item{scoring_rule}{A character string specifying the scoring rule to use
as input}
}
\value{
An error is raised if any of the inputs is not valid.
}
\description{
Validate metric selection for output types
}
\examples{
\dontrun{
# Invalid combinations
check_metric_selection("mean", "wis") # Raises an error
check_metric_selection("quantile", "ae_point") # Raises an error
check_metric_selection("mean", "se_point") # Returns TRUE
check_metric_selection("quantile", "wis") # Returns TRUE
}
}
