% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-s3.R
\name{plot.importance_summary}
\alias{plot.importance_summary}
\title{Plot method for overall importance summary}
\usage{
\method{plot}{importance_summary}(x, ...)
}
\arguments{
\item{x}{An object of class \code{importance_summary}.}

\item{...}{Additional arguments passed to the plot method.}
}
\description{
Plot method for overall importance summary
}
