% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-s3.R
\name{plot.model_imp_tbl}
\alias{plot.model_imp_tbl}
\title{Plot method for model importance score table}
\usage{
\method{plot}{model_imp_tbl}(x, ...)
}
\arguments{
\item{x}{An object of class \code{model_imp_tbl}.}

\item{...}{Additional arguments passed to the plot method.}
}
\description{
Plot method for model importance score table
}
