% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-s3.R
\name{print.importance_summary}
\alias{print.importance_summary}
\title{Print method for overall importance summary}
\usage{
\method{print}{importance_summary}(x, ...)
}
\arguments{
\item{x}{An object of class \code{importance_summary}.}

\item{...}{Additional arguments passed to the print method.}
}
\description{
Print method for overall importance summary
}
