% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-s3.R
\name{summary.importance_summary}
\alias{summary.importance_summary}
\title{Summary method for overall importance summary}
\usage{
\method{summary}{importance_summary}(object, ...)
}
\arguments{
\item{object}{An object of class \code{importance_summary}.}

\item{...}{Additional arguments passed to the summary method.}
}
\description{
Summary method for overall importance summary
}
