% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{target_data_example}
\alias{target_data_example}
\title{Example target data for modelimportance article vignette}
\format{
A data frame with 8 rows and 4 columns:
\describe{
\item{target_end_date}{Date. Date of the observation.}
\item{target}{Character. Target name.}
\item{location}{Character. FIPS code for the US location.}
\item{observation}{Numeric. Observed value.}
}
}
\source{
Sourced from \code{hubExamples::forecast_target_ts}
(hubverse-org/hubExamples v1.0.0).
}
\usage{
target_data_example
}
\description{
A subset of \code{hubExamples::forecast_target_ts} matching the locations and
target end dates in \code{forecast_data_example}.
}
\keyword{datasets}
