% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{target_data_ma}
\alias{target_data_ma}
\title{Target data for Massachusetts used in vignette runtime data}
\format{
A data frame with 1 row and 4 columns:
\describe{
\item{target_end_date}{Date. Date of the observation.}
\item{target}{Character. Target name.}
\item{location}{Character. FIPS code for the US location.}
\item{observation}{Numeric. Observed value.}
}
}
\source{
Sourced from \code{hubExamples::forecast_target_ts}
(hubverse-org/hubExamples v1.0.0).
}
\usage{
target_data_ma
}
\description{
A subset of \code{hubExamples::forecast_target_ts} for location "25"
(Massachusetts) and target end date 2022-12-24.
}
\keyword{datasets}
