% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid_input_data.R
\name{valid_input_data}
\alias{valid_input_data}
\title{Check the input forecast data structure}
\usage{
valid_input_data(forecast_data)
}
\arguments{
\item{forecast_data}{A data.frame with the predictions that is or can be
coerced to a model_out_tbl format.}
}
\value{
a model_out_tbl format that has a single output type
}
\description{
Check the input forecast data structure
}
\examples{
\dontrun{
library(hubUtils)
library(dplyr)
hub_path <- system.file("testhubs/flusight", package = "hubUtils")
hub_con <- connect_hub(hub_path)
hub_con |>
  filter(output_type == "quantile") |>
  collect() |>
  valid_input_data()
  }
}
