% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_input_data.R
\name{validate_input_data}
\alias{validate_input_data}
\title{Check the input forecast data structure}
\usage{
validate_input_data(forecast_data, target_data)
}
\arguments{
\item{forecast_data}{A data.frame with the predictions that is or can be
coerced to a model_out_tbl format.}

\item{target_data}{A data.frame with the target values. This data must follow
the oracle output format.}
}
\value{
a model_out_tbl format that has a single output type
}
\description{
Check the input forecast data structure
}
\examples{
\dontrun{
library(hubUtils)
library(dplyr)
hub_path <- system.file("testhubs/flusight", package = "hubUtils")
hub_con <- connect_hub(hub_path)
hub_con |>
  filter(output_type == "quantile") |>
  collect() |>
  validate_input_data()
}
}
