% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-s3.R
\name{print.model_imp_tbl}
\alias{print.model_imp_tbl}
\title{Print method for model importance score table}
\usage{
\method{print}{model_imp_tbl}(x, ...)
}
\arguments{
\item{x}{An object of class \code{model_imp_tbl}.}

\item{...}{Additional arguments passed to the print method.}
}
\description{
Print method for model importance score table
}
