% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_data_by_task.R
\name{split_data_by_task}
\alias{split_data_by_task}
\title{Split the input data by a single task.}
\usage{
split_data_by_task(valid_tbl, weighted, training_window_length)
}
\arguments{
\item{valid_tbl}{A data.frame containing forecast data and target data,
processed by the function `validate_input_data()`.}

\item{weighted}{Boolean indicating whether model weighting should be done.
If `FALSE`, all models are given equal weight.
If `TRUE`, model weights are estimated.}

\item{training_window_length}{An integer value representing the time interval
of historical data used during the training process
to estimate model weights.
Default is `0`, meaning that no prior data is available for training.}
}
\value{
A list of data sets, each corresponding to a single task.
}
\description{
This function splits the input data by a single task with a
combination of horizon, location, and target_end_date
}
\examples{
\dontrun{
library(dplyr)
library(purrr)
library(hubExamples)
forecast_data <- hubExamples::forecast_outputs |>
  dplyr::filter(
    output_type \%in\% c("mean"),
    location \%in\% c("25", "48"),
    horizon \%in\% c(1, 2)
  )
target_data <- hubExamples::forecast_target_ts |>
  dplyr::filter(
    date \%in\% unique(forecast_data$target_end_date),
    location == "25"
  ) |>
  # Rename columns to match the oracle output format
  rename(
    target_end_date = date,
    oracle_value = observation
  )
valid_tbl <- validate_input_data(forecast_data, target_data)
split_data_by_task(valid_tbl, weighted = FALSE, training_window_length = 0)
split_data_by_task(valid_tbl, weighted = TRUE, training_window_length = 1)
}
}
