% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_data_by_task.R
\name{split_data_by_task}
\alias{split_data_by_task}
\title{Split forecast data by a single task}
\usage{
split_data_by_task(valid_tbl)
}
\arguments{
\item{valid_tbl}{A data.frame in a model_out_tbl format.}
}
\value{
A list of data sets, each corresponding to a single task.
}
\description{
Split forecast data by a single task
}
\examples{
\dontrun{
library(dplyr)
library(hubUtils)
hub_path <- system.file("testhubs/flusight", package = "hubUtils")
hub_con <- connect_hub(hub_path)
hub_con |>
  filter(output_type == "quantile") |>
  collect() |>
  valid_input_data() |>
  split_data_by_task()
}
}
