% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_data_by_task.R
\name{split_data_by_task}
\alias{split_data_by_task}
\title{Split the input data by a single task and make a list of data sets}
\usage{
split_data_by_task(valid_tbl)
}
\arguments{
\item{valid_tbl}{A data.frame containing forecast data and target data,
processed by the function \code{validate_input_data()}.}
}
\value{
A list of data sets, each corresponding to a single task.
}
\description{
This function splits the input data by a single task with a
combination of task IDs and returns a list of data sets, each corresponding
to a single task.
}
