% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-s3.R
\name{summary.model_imp_tbl}
\alias{summary.model_imp_tbl}
\title{Summary method for model importance score table}
\usage{
\method{summary}{model_imp_tbl}(object, ...)
}
\arguments{
\item{object}{An object of class \code{model_imp_tbl}.}

\item{...}{Additional arguments passed to the summary method.}
}
\description{
Summary method for model importance score table
}
