% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_input_data.R
\name{validate_input_data}
\alias{validate_input_data}
\title{Check the input forecast data structure and validate it.}
\usage{
validate_input_data(forecast_data, oracle_output_data)
}
\arguments{
\item{forecast_data}{A data.frame with the predictions that is or can be
coerced to a model_out_tbl format.}

\item{oracle_output_data}{A data.frame with the target values.
This data must follow the oracle output format.}
}
\value{
a model_out_tbl format for forecast data
}
\description{
Check the input forecast data structure and validate it.
}
\examples{
\dontrun{
library(dplyr)
library(hubExamples)
forecast_data <- hubExamples::forecast_outputs |>
  dplyr::filter(
    output_type \%in\% c("quantile"),
    location == "25",
    horizon == 1
  )
target_data <- hubExamples::forecast_target_ts |>
  dplyr::filter(
    date \%in\% unique(forecast_data$target_end_date),
    location == "25"
  ) |>
  # Rename columns to match the oracle output format
  rename(
    target_end_date = date,
    oracle_value = observation
  )
}
}
