% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_input_data.R
\name{validate_input_data}
\alias{validate_input_data}
\title{Check the input forecast data structure and validate it.}
\usage{
validate_input_data(forecast_data, oracle_output_data)
}
\arguments{
\item{forecast_data}{A data.frame with the predictions that is or can be
coerced to a model_out_tbl format.}

\item{oracle_output_data}{A data.frame with the target values.
This data must follow the oracle output format.}
}
\value{
a model_out_tbl format for forecast data
}
\description{
Check the input forecast data structure and validate it.
}
